service/data-ingester/code/requirements.txt
attrs==25.4.0
autobahn==19.11.2
Automat==25.4.16
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
constantly==23.10.4
cryptography==46.0.3
Flask==3.1.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiteconnect==5.0.1
MarkupSafe==3.0.2
numpy==2.0.2
onetimepass==1.0.1
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
schedule==1.2.2
service-identity==24.2.0
tomli==2.3.0
Twisted==25.5.0
txaio==23.6.1
tzdata==2025.2
urllib3==2.5.0
mysql-connector-python==9.4.0
Flask-Limiter